+++ mfhdf/test/hdftest.c
@@ -480,24 +480,24 @@ main(void)
     CHECK(status, FAIL, "SDreaddata");
 
     /* verify the data values retrieved from 'DataSetAlpha' */
-    if (data[0] != -17.5) {
-        fprintf(stderr, "Wrong value returned loc 0: %f\n", (float)data[0]);
+    if (data[0] != -17.5F) {
+        fprintf(stderr, "Wrong value returned loc 0: %f\n", (double)data[0]);
         num_errs++;
     }
-    if (data[3] != -17.5) {
-        fprintf(stderr, "Wrong value returned loc 3: %f\n", (float)data[3]);
+    if (data[3] != -17.5F) {
+        fprintf(stderr, "Wrong value returned loc 3: %f\n", (double)data[3]);
         num_errs++;
     }
-    if (data[5] != 1.0) {
-        fprintf(stderr, "Wrong value returned loc 5: %f\n", (float)data[5]);
+    if (data[5] != 1.0F) {
+        fprintf(stderr, "Wrong value returned loc 5: %f\n", (double)data[5]);
         num_errs++;
     }
-    if (data[6] != -17.5) {
-        fprintf(stderr, "Wrong value returned loc 6: %f\n", (float)data[6]);
+    if (data[6] != -17.5F) {
+        fprintf(stderr, "Wrong value returned loc 6: %f\n", (double)data[6]);
         num_errs++;
     }
-    if (data[8] != 4.0) {
-        fprintf(stderr, "Wrong value returned loc 8: %f\n", (float)data[8]);
+    if (data[8] != 4.0F) {
+        fprintf(stderr, "Wrong value returned loc 8: %f\n", (double)data[8]);
         num_errs++;
     }
